# server/openapi.yaml
openapi: 3.1.0
info:
  title: Read Your Bible Through API
  version: 2.0.0
  description: >
    Phase 0 contract. Token-authenticated API. All timestamps (`updatedAt`) are
    ISO-8601 UTC strings (e.g. `2026-06-29T12:00:00.000Z`) so lexicographic order
    equals chronological order (required for last-write-wins).
servers:
  - url: '{baseUrl}/api'
    variables:
      baseUrl:
        default: http://localhost:1701
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PassageTrack:
      type: object
      required: [passageDate, passageType, isRead, updatedAt]
      properties:
        id: {type: string}
        passageDate: {type: string, description: 'YYYY-MM-DD'}
        passageType: {type: string, enum: [bible, proverbs, psalms]}
        isRead: {type: boolean}
        trackDate: {type: string}
        updatedAt: {type: string, description: ISO-8601 UTC client action-time}
    ChapterTrack:
      type: object
      required: [book, chapter, isRead, updatedAt]
      properties:
        id: {type: string}
        book: {type: string, description: USFM id (GEN, PRO, PSA, ...)}
        chapter: {type: string}
        isRead: {type: boolean}
        trackDate: {type: string}
        updatedAt: {type: string, description: ISO-8601 UTC client action-time}
    TracksResponse:
      type: object
      properties:
        passages: {type: array, items: {$ref: '#/components/schemas/PassageTrack'}}
        chapters: {type: array, items: {$ref: '#/components/schemas/ChapterTrack'}}
security:
  - bearerAuth: []
paths:
...
  /user/tracks:
    get:
      summary: Incremental pull of tracks for the authenticated user.
      parameters:
        - name: since
          in: query
          required: false
          schema: {type: string}
          description: ISO-8601 UTC. Returns rows with updatedAt > since. Omit for full pull.
      responses:
        '200':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TracksResponse'}
        '401': {description: Missing/invalid token}
    put:
      summary: Batch upsert (outbox drain). Idempotent, last-write-wins by updatedAt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                passages:
                  type: array
                  items: {$ref: '#/components/schemas/PassageTrack'}
                chapters:
                  type: array
                  items: {$ref: '#/components/schemas/ChapterTrack'}
      responses:
        '200':
          description: Resolved server state for upserted keys.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TracksResponse'}
        '401': {description: Missing/invalid token}